+++ CHANGELOG.md
@@ -16,6 +16,7 @@
   - Generic Type validator - verify that the 'fromVersion' field is above 6.5.0
   - Generic Module validator - verify that the 'fromVersion' field is above 6.5.0
   - Generic Definition validator - verify that the 'fromVersion' field is above 6.5.0
+ * Expanded Format command to support Generic Objects - Fixes generic objects according to their validations.
 
 # 1.4.5
 * Enhanced the **postman-codegen** command to name all generated arguments with lower case.